refactor: images, upload controller
